Allow building and publishing of docs on dev
`MIX_ENV=docs` was needed to build and publish docs. I moved it to the
dev environment so we wouldn't need a special environment for tests. I
also added `runtime: false` to make sure that it wouldn't do anything
when starting the application. This also saves time by not having to
recompile just to produces documentation. If dev is already compiled you
don't have to wait again in a new environment.

Amos King @adkron <amos@binarynoggin.com>