Feature/property test (#5)
* Adds a single property test

The property test is to check that a jpeg is equal to the same jpeg
after it is turned into yuv and back to jpeg. Currently I only know how
to get the hearders to be the same. The images look the same, but have
slightly different binary.

* Start working on a property test for solid colors

* Testing out some ranges of generating numbers

I was looking at the aggregates of the images properties and wanted to
try to get some realistic numbers. This is a first pass.

* Write test to check nif computation time.

Set the functions to be dirty cpu bound so they don't
block regular schedulers

* Removes verbose in property tests

Verbose creates a lot of output by default.

Co-authored-by: Amos King <amos@binarynoggin.com> and Connor Rigby <connor@binarynoggin.com>