Document types in Turbojpeg
Documenting is important and types are important to the external interface of
the library. To be able to document I had to add a wrapper to the Native code. I
think that is okay for us to add a layer in front of the native for a little
more control.

I decided to capture the exceptions coming from the native code because I'm not
sure if the Elixir processes using this should die if the native code has
issues. Not the choice is in the hands of the user.