Adds Circle CI
There was a lot of churn in getting the build setup. I'm squashing it because none of it makes sense alone except for some README updates.

* Adds dev deps to README

In order to run tests we need imagemagick so I added to the dev instructions to
add imagemagick to the system.

* Add a CircleCI build

We want to have an automated build to run tests.

* Adds CircleCI badge

It is always nice to let people know our build status.

* Fix circle file format

We can't nest under version. The error message said something about not being
able to map values so it wasn't clear what was wrong.

* Fixes installing of libturbojpeg for the build

We need to do an apt-get update before we can install libturbojpeg. If we don't
do that then apt-get can't find it.

* Fixes Ubuntu package name for libturbojpeg

* Stops apt-get prompts on the build

The build hangs when installing deps with apt-get because apt-get want to ask if
you really want to install. Adding a `-y` switch should do the trick.

* Requires git for circle ci

Right now we have a dependency on mogrify from Connor's fork. Once his PR is
merged we will no longer need git to be installed and can update the dependency.

NOTE: This dependency is for testing only

* Fixes 404s in circle ci setup

Apt-get was getting 404s the recommended course of action was apt-get update.

* Changes libturbojpeg ubuntu install

We need the dev package of libturbojpeg in order to build.

* Installs all the libturbojpeg packages I can find

This is getting frustrating, but it appears that the build and my local install
are doing different things and I can't figure it out at this moment.

* Attempting different jpeg install for ubuntu

I can't seem to find the right one, but apt-get search came up with this so I'm
trying it.

* Use latest Elixir builder

We believe the libturbojpeg0-dev package should be there.

* Updates the deps

There is an error in the version of Shmex and Bundlex we were using that
presented itself in elixir 1.10. Updating it fixes the issue.

* Adds libturbojpeg

We need the dev package and the normal one to build.

* Add pkg-config for ci

This is needed to build schmex

* Changes build's MIX_ENV to test

The build was set to prod which wouldn't install propcheck. Changing it should
be able to run the tests.

* Removes performance test

We never did get the test right, and it was just for local playing any way.